requirements.txt
Django>=4.0,<5.0
psycopg2-binary==2.9.7
djangorestframework==3.15.2
gunicorn>=20.0
django-cors-headers==4.6.0
django-extensions
django-sslserver